chore: move logger initialization to a service
